+++ halo_itsm/CHANGELOG.md
@@ -1,3 +1,10 @@
+# 2.0.0-beta
+## Added 
+- Added support for ticket types in configuration parameters.
+- Added support for mapping custom status fields.
+- Added support for dynamic queue mappings based on the ticket type.
+- Added resolutions for error logs.
+
 # 1.0.0
 ## Added
 - Initial release.